Add download button
The download button is the submit button of a form, the rest of which
is hidden and contains all the info on the primer designs.
The download route produces a csv text string and return a text/csv
response.

The create_file_for_download function is no longer required so
app/helpers.py has been removed as well as the test for it.